Mo_LatestDesignChanges_v1
Hey Guys, I started on working with the design finesse a bit. I still
have a ways to go, but there were a few things as a result of the
design changes that got skewed. One was the highlighted post. I change
the height of it, so now the ing is a little off centered. Could
someone fix this.